playerservice/requirements.txt
Django>=4.0,<5.0
Pillow>=9.0,<10.0
djangorestframework>=3.14,<4.0
pytest>=7.0             
pytest-django>=4.5     
locust>=2.15              

psycopg2-binary==2.9.9